[Refator] mockUser 파일 분리
