+++ src/components/mockData/mockUser.jsx
@@ -1,0 +1,40 @@
+const mockUser = [
+  {
+    name: '이강혁',
+    studentId: '1234567',
+    tel: '01012345678',
+    department: 'SW',
+    email: '123@123.123',
+    cardinal: 1,
+    position: 'BE',
+  },
+  {
+    name: '박예진',
+    studentId: '1234567',
+    tel: '01012345678',
+    department: 'SW',
+    email: '123@123.123',
+    cardinal: 2,
+    position: 'FE',
+  },
+  {
+    name: '조예진',
+    studentId: '1234567',
+    tel: '01012345678',
+    department: 'SW',
+    email: '123@123.123',
+    cardinal: 3,
+    position: 'D',
+  },
+  {
+    name: '김성민',
+    studentId: '1234567',
+    tel: '01012345678',
+    department: 'SW',
+    email: '123@123.123',
+    cardinal: 3,
+    position: 'MA',
+  },
+];
+
+export default mockUser;
