+++ src/pages/User.jsx
@@ -2,6 +2,7 @@ import styled from 'styled-components';
 import UserHeader from '../components/Header/UserHeader';
 import Generation from '../components/Header/Generation';
 import UserName from '../components/UserName';
+import mockUser from '../components/mockData/mockUser';
 
 const StyledMember = styled.div`
   width: 370px;
@@ -15,49 +16,10 @@ const MemberList = styled.div`
   height: 100%;
 `;
 
-const mockUser = [
-  {
-    name: '이강혁',
-    studentId: '1234567',
-    tel: '01012345678',
-    department: 'SW',
-    email: '123@123.123',
-    cardinal: 1,
-    position: 'BE',
-  },
-  {
-    name: '박예진',
-    studentId: '1234567',
-    tel: '01012345678',
-    department: 'SW',
-    email: '123@123.123',
-    cardinal: 2,
-    position: 'FE',
-  },
-  {
-    name: '조예진',
-    studentId: '1234567',
-    tel: '01012345678',
-    department: 'SW',
-    email: '123@123.123',
-    cardinal: 3,
-    position: 'D',
-  },
-  {
-    name: '김성민',
-    studentId: '1234567',
-    tel: '01012345678',
-    department: 'SW',
-    email: '123@123.123',
-    cardinal: 3,
-    position: 'MA',
-  },
-];
-
 const Member = () => {
   return (
     <StyledMember>
-      <UserHeader mockUser={mockUser} />
+      <UserHeader />
       <Generation />
       <MemberList>
         {mockUser.map((user) => (
